requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.7.7
    # via flask-migrate
apispec[yaml]==5.1.1
    # via
    #   -r requirements.in
    #   apispec-webframeworks
    #   flask-apispec
apispec-webframeworks==0.5.2
    # via -r requirements.in
arrow==1.2.2
    # via -r requirements.in
attrs==21.4.0
    # via pytest
backcall==0.2.0
    # via ipython
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   flask
    #   flask-shell-ipython
colorlog==6.6.0
    # via seveno-pyutil
decorator==5.1.1
    # via ipython
et-xmlfile==1.1.0
    # via openpyxl
factory-boy==3.2.1
    # via -r requirements.in
faker==13.3.2
    # via
    #   -r requirements.in
    #   factory-boy
flask==2.0.3
    # via
    #   -r requirements.in
    #   flask-apispec
    #   flask-cors
    #   flask-migrate
    #   flask-shell-ipython
    #   flask-sqlalchemy
flask-apispec==0.11.0
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-shell-ipython==0.4.1
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
geoalchemy2==0.11.1
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
infinity==1.5
    # via intervals
iniconfig==1.1.1
    # via pytest
intervals==0.9.2
    # via -r requirements.in
ipython==7.27.0
    # via
    #   -r requirements.in
    #   flask-shell-ipython
itsdangerous==2.1.1
    # via
    #   -r requirements.in
    #   flask
jedi==0.18.1
    # via ipython
jinja2==3.0.3
    # via flask
lorem==0.1.1
    # via -r requirements.in
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.14.1
    # via
    #   -r requirements.in
    #   flask-apispec
    #   marshmallow-enum
    #   seveno-pyutil
    #   webargs
marshmallow-enum==1.5.1
    # via -r requirements.in
matplotlib-inline==0.1.3
    # via ipython
numpy==1.22.3
    # via pandas
openpyxl==3.0.9
    # via -r requirements.in
packaging==21.3
    # via
    #   geoalchemy2
    #   pytest
    #   webargs
pandas==1.4.1
    # via -r requirements.in
parso==0.8.3
    # via jedi
passlib==1.7.4
    # via -r requirements.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==1.0.0
    # via pytest
prettytable==3.2.0
    # via -r requirements.in
prompt-toolkit==3.0.28
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
py==1.11.0
    # via pytest
pygments==2.11.2
    # via
    #   ipython
    #   seveno-pyutil
pyjwt==2.3.0
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pytest==7.1.0
    # via
    #   -r requirements.in
    #   pytest-mock
pytest-mock==3.7.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   arrow
    #   faker
    #   pandas
    #   seveno-pyutil
python-dotenv==0.19.2
    # via -r requirements.in
python-http-client==3.3.7
    # via sendgrid
pytz==2021.3
    # via
    #   -r requirements.in
    #   pandas
    #   seveno-pyutil
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via apispec
requests==2.27.1
    # via -r requirements.in
sendgrid==6.9.7
    # via -r requirements.in
seveno-pyutil==0.5.8
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-cors
    #   python-dateutil
    #   sqlalchemy-utils
sqlalchemy==1.4.32
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
    #   geoalchemy2
    #   sqlalchemy-utils
sqlalchemy-utils==0.38.2
    # via -r requirements.in
sqlparse==0.4.2
    # via
    #   -r requirements.in
    #   seveno-pyutil
starkbank-ecdsa==2.0.3
    # via sendgrid
tomli==2.0.1
    # via pytest
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
tzdata==2021.5
    # via pytz-deprecation-shim
tzlocal==4.1
    # via
    #   -r requirements.in
    #   seveno-pyutil
unipath==1.1
    # via -r requirements.in
urllib3==1.26.8
    # via requests
wcwidth==0.2.5
    # via
    #   prettytable
    #   prompt-toolkit
webargs==8.1.0
    # via flask-apispec
werkzeug==2.0.3
    # via flask
xlsxwriter==3.0.3
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
